+++ packages/react-core/src/components/Divider/Divider.tsx
@@ -13,18 +13,21 @@ export interface DividerProps extends React.HTMLProps<HTMLElement> {
   className?: string;
   /** The component type to use */
   component?: 'hr' | 'li' | 'div';
+  /** Flag to indicate the divider is vertical (must be in a flex layout) */
+  isVertical?: boolean;
 }
 
 export const Divider: React.FunctionComponent<DividerProps> = ({
   className = '',
   component = DividerVariant.hr,
+  isVertical = false,
   ...props
 }: DividerProps) => {
   const Component: any = component;
 
   return (
     <Component
-      className={css(styles.divider, className)}
+      className={css(styles.divider, isVertical && styles.modifiers.vertical, className)}
       {...(component !== 'hr' && { role: 'separator' })}
       {...props}
     />