+++ packages/react-core/src/components/Modal/ModalBoxHeader.tsx
@@ -1,29 +1,21 @@
 import * as React from 'react';
 
-import { Title, TitleSizes } from '../Title';
+import styles from '@patternfly/react-styles/css/components/ModalBox/modal-box';
+import { css } from '@patternfly/react-styles';
 
 export interface ModalBoxHeaderProps {
   /** Content rendered inside the Header */
   children?: React.ReactNode;
   /** Additional classes added to the button */
   className?: string;
-  /** Flag to hide the title */
-  hideTitle?: boolean;
-  /** The heading level to use */
-  headingLevel?: 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6';
 }
 
 export const ModalBoxHeader: React.FunctionComponent<ModalBoxHeaderProps> = ({
   children = null,
   className = '',
-  hideTitle = false,
-  headingLevel = 'h1',
   ...props
-}: ModalBoxHeaderProps) =>
-  hideTitle ? null : (
-    <React.Fragment>
-      <Title size={TitleSizes['2xl']} headingLevel={headingLevel} className={className} {...props}>
-        {children}
-      </Title>
-    </React.Fragment>
-  );
+}: ModalBoxHeaderProps) => (
+  <div className={css(styles.modalBoxTitle, className)} {...props}>
+    {children}
+  </div>
+);