+++ packages/react-core/src/components/Modal/__tests__/ModalBoxHeader.test.tsx
@@ -7,13 +7,3 @@ test('ModalBoxHeader Test', () => {
   const view = shallow(<ModalBoxHeader>This is a ModalBox header</ModalBoxHeader>);
   expect(view).toMatchSnapshot();
 });
-
-test('ModalBoxHeader Test with H3', () => {
-  const view = shallow(<ModalBoxHeader headingLevel="h3">This is a ModalBox header</ModalBoxHeader>);
-  expect(view).toMatchSnapshot();
-});
-
-test('ModalBoxHeader Test hideTitle', () => {
-  const view = shallow(<ModalBoxHeader hideTitle>This is a ModalBox header</ModalBoxHeader>);
-  expect(view).toMatchSnapshot();
-});