+++ packages/react-core/src/components/Toolbar/ToolbarItem.tsx
@@ -4,6 +4,7 @@ import { css } from '@patternfly/react-styles';
 
 import { ToolbarBreakpointMod } from './ToolbarUtils';
 import { formatBreakpointMods, toCamel } from '../../helpers/util';
+import { Divider } from '../Divider';
 
 export enum ToolbarItemVariant {
   separator = 'separator',
@@ -18,7 +19,7 @@ export enum ToolbarItemVariant {
 export interface ToolbarItemProps extends React.HTMLProps<HTMLDivElement> {
   /** Classes applied to root element of the data toolbar item */
   className?: string;
-  /** TODO: Support 'separator' element as a <Divider component="div" variant="horizontal" />. A type modifier which modifies spacing specifically depending on the type of item */
+  /** A type modifier which modifies spacing specifically depending on the type of item */
   variant?:
     | ToolbarItemVariant
     | 'bulk-select'
@@ -26,7 +27,8 @@ export interface ToolbarItemProps extends React.HTMLProps<HTMLDivElement> {
     | 'pagination'
     | 'search-filter'
     | 'label'
-    | 'chip-group';
+    | 'chip-group'
+    | 'separator';
   /** An array of objects representing the various modifiers to apply to the data toolbar item at various breakpoints */
   breakpointMods?: ToolbarBreakpointMod[];
   /** id for this data toolbar item */
@@ -42,21 +44,27 @@ export const ToolbarItem: React.FunctionComponent<ToolbarItemProps> = ({
   id,
   children,
   ...props
-}: ToolbarItemProps) => (
-  <div
-    className={css(
-      styles.toolbarItem,
-      variant &&
-        styles.modifiers[
-          toCamel(variant) as 'bulkSelect' | 'overflowMenu' | 'pagination' | 'searchFilter' | 'label' | 'chipGroup'
-        ],
-      formatBreakpointMods(breakpointMods, styles),
-      className
-    )}
-    {...(variant === 'label' && { 'aria-hidden': true })}
-    id={id}
-    {...props}
-  >
-    {children}
-  </div>
-);
+}: ToolbarItemProps) => {
+  if (variant === ToolbarItemVariant.separator) {
+    return <Divider className={css(styles.modifiers.vertical, className)} {...props} />;
+  }
+
+  return (
+    <div
+      className={css(
+        styles.toolbarItem,
+        variant &&
+          styles.modifiers[
+            toCamel(variant) as 'bulkSelect' | 'overflowMenu' | 'pagination' | 'searchFilter' | 'label' | 'chipGroup'
+          ],
+        formatBreakpointMods(breakpointMods, styles),
+        className
+      )}
+      {...(variant === 'label' && { 'aria-hidden': true })}
+      id={id}
+      {...props}
+    >
+      {children}
+    </div>
+  );
+};