+++ packages/react-inline-edit-extension/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.3.23](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.3.22...@patternfly/react-inline-edit-extension@4.3.23) (2020-05-07)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.3.22](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.3.21...@patternfly/react-inline-edit-extension@4.3.22) (2020-05-07)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.3.21](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.3.20...@patternfly/react-inline-edit-extension@4.3.21) (2020-05-06)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.3.20](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.3.19...@patternfly/react-inline-edit-extension@4.3.20) (2020-05-06)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.3.19](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.3.18...@patternfly/react-inline-edit-extension@4.3.19) (2020-05-05)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.3.18](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.3.17...@patternfly/react-inline-edit-extension@4.3.18) (2020-05-05)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.3.17](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.3.16...@patternfly/react-inline-edit-extension@4.3.17) (2020-05-05)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
 ## [4.3.16](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.3.15...@patternfly/react-inline-edit-extension@4.3.16) (2020-05-04)
 
 **Note:** Version bump only for package @patternfly/react-inline-edit-extension