+++ packages/react-topology/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.23](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@4.2.22...@patternfly/react-topology@4.2.23) (2020-05-07)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [4.2.22](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@4.2.21...@patternfly/react-topology@4.2.22) (2020-05-07)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [4.2.21](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@4.2.20...@patternfly/react-topology@4.2.21) (2020-05-06)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [4.2.20](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@4.2.19...@patternfly/react-topology@4.2.20) (2020-05-06)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [4.2.19](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@4.2.18...@patternfly/react-topology@4.2.19) (2020-05-05)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [4.2.18](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@4.2.17...@patternfly/react-topology@4.2.18) (2020-05-05)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [4.2.17](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@4.2.16...@patternfly/react-topology@4.2.17) (2020-05-05)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
 ## [4.2.16](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@4.2.15...@patternfly/react-topology@4.2.16) (2020-05-04)
 
 **Note:** Version bump only for package @patternfly/react-topology